Moved more create tests to creation test module
